:sparkles: Add .editorconfig and .gitignore files
Added .editorconfig file to set the coding style and formatting rules for the project. The file specifies the use of spaces for indentation, sets the indent size to 2 spaces, uses LF (line feed) as the end-of-line character, and sets the character encoding to UTF-8. It also ensures that a final newline is always inserted at the end of the file.